Refactor Layout to Xilems/Flutters BoxConstraints (#5)
I did some exploration and came back to the conclusion to use BoxConstraints for a base layouting protocol, used in Flutter and Jetpack Compose.
(After trying the SwiftUI layout protocol which, it seems is not as powerful as BoxConstraints,
SwiftUI seems to do internal layout magic, that isn't really documented).

BoxConstraints seem to work well in Flutter and Jetpack Compose and are simple to understand.

This cop... ehhm adapts the layouting system to BoxConstraints as used in Xilem.
(I think it is time to attribute where most of the code comes from (Adding license and mentions etc.))
This shifts a little bit towards being *the* xilem TUI implementation (e.g. use kurbo etc. basically the same base as Xilem).
But it remains to be seen whether it diverges in the future (e.g. styling or layouting etc.).

It changes the layouting behavior currently (by "hugging"). I expect to add more layout modifiers soon to control the layout behavior.

The `WrappedText` widget doesn't work anymore, this needs some bigger refactor and rethinking for styling text in general (it was never meant to be more than just a quick'n dirty experiment anyway).