Add `WeightedLinearLayout` View/Widget
This forces all children to have a weighted size,
instead of giving the freedom to choose their size via constraints.
By default a view has a weight of 1.0, this can be
overridden via the `WeightedLayoutElement` (or `view.weight(f64)`)

One possibly interesting pattern here is the dynamic
cast of the type (`WeightedLayoutElement`) and when
it fails a default of 1.0 is chosen.

This pattern could be used in other contexts as well
to avoid having to wrap all views in context-specific
types such as the `WeightedLayoutElement`.

There are (slight) disadvantages though,
One is that it doesn't cascade.
So something like `view.weight(1.5).other_modifier()`
does have no effect (or rather has a default weight of 1.0) in the
`WeightedLinearLayout` context.
It can also be used in different contexts where it doesn't make sense.
So it would be interesting to explore more statically enforced solutions
still (while not unnecessarily increasing type complexity)