Fix print_buffer (#12)
* Fix print_buffer

It's still not perfect, but I think it's pretty useful already.
A nuisance is that the test output starts immediately after the last character
in the last line of the widget.
I haven't found a way to prevent that.

* Ensure errors from the spawned thread are not masked

fixes #13

The error with multiple render_view() tests was caused by global logging being
initialized more than once.
See also https://github.com/tokio-rs/console/issues/505

My workaround is a bit hacky, but I think there are a bunch of things, including
logging, that should be set up using a configuation struct passed to `fn new()`
anyway. Cleaning that workaround up would go with that task automatically.