Animatable directly in view (#16)
This sketches out a potential animation API (that is very likely too happen in a similar form as this).

Currently values can be animated via the new `Animatable` trait which is basically a `View` trait, but (currently) without `AppState` for simplictity (will likely come in the future, to e.g. change state after an animation has finished etc.), and the value that is animated is passed via generic parameter instead of an associated type (in contrast to `View::Element`).

Additionally it adds a `Tweenable` trait, which is also very similar to the `View` trait, but takes a `ratio: f64` in the rebuild/build methods to interpolate be"tween" values.
This adds the following animatables:

* `Tween` which plays with a(n animatable) `play_speed` and a duration a `Tweenable`
* `Lerp` which linearly interpolates a `Tweenable` with an animatable ratio
* `LowPassIIR` which is a just a IIR low pass filter that filters an animatable `f64` with a decay (which is not yet animatable but could be too)
* literal/const values (currently implemented for `f64` and `u32`) are also animatables

Currently a "base" Tweenable is a `Range<impl Animatable<f64>>`.
This can be modified with ease functions (which in turn are `Tweenable`s as well).
The following functions are currently defined:

* `Reverse` - just reverses the `Tweenable`
* `QuadraticIn`
* `QuadraticOut`
* `QuadraticInOut`
* `ElasticInOut`

Plan is to be able to compose Tweenables like a sequence via tuples: 
`impl Tweenable for (impl Tweenable, impl Tweenable)`.

Additionally it adds an example showing all of the above.

This very clearly moves the idea of modifier based styling forward, as all of the above is basically modifier based animation (e.g. `(0.0..1.0).reverse().quadratic_in().elastic_in_out_ease().tween(..)`).