Refactored Animatable into the element tree and fixed some bugs (#18)
Having Animatables in the view has some issues, e.g. memoize doesn't work.
(It was never planned to keep it there anyway, but for simplicity it was done like this for now)
It also requires for each animation update to rerender the whole view.
(This happens currently non-the-less, but can be changed in the future)

This moves the Animatable and Tweenable logic into the element tree, and adds a new ChangeFlag `ANIMATION`.
`Animatable` and `Tweenable` slowly converge back to `View` traits And can thus be genereated as such to have support for other core views like memoize etc. :).

There were also a few issues with pods not being marked (`Pod::mark`) in the owner views